shared/scripts/requirements.txt
faker>=24.0
psycopg2-binary>=2.9
pymongo>=4.6
tqdm>=4.66
boto3>=1.34
